+++ tests/test_response_validation/unit/response_tester/test_dict.py
@@ -23,7 +23,7 @@ def test_valid_dict() -> None:
     """
     Asserts that valid data passes successfully.
     """
-    tester.test_dict(schema=schema, data=data, parent='placeholder')
+    tester.test_dict(schema=schema, data=data, reference='placeholder')
 
 
 def test_bad_data_type() -> None:
@@ -33,7 +33,7 @@ def test_bad_data_type() -> None:
     with pytest.raises(
         SwaggerDocumentationError, match="Expected response to be <class 'dict'> but found <class 'list'>."
     ):
-        tester.test_dict(schema=schema, data=[data], parent='placeholder')
+        tester.test_dict(schema=schema, data=[data], reference='placeholder')
 
 
 def test_unmatched_lengths() -> None:
@@ -45,7 +45,7 @@ def test_unmatched_lengths() -> None:
         SwaggerDocumentationError,
         match='The following properties seem to be missing from your OpenAPI/Swagger documentation: `extra key`',
     ):
-        tester.test_dict(schema=schema, data=weird_data, parent='placeholder')
+        tester.test_dict(schema=schema, data=weird_data, reference='placeholder')
 
 
 def test_schema_key_not_in_response():
@@ -56,7 +56,7 @@ def test_schema_key_not_in_response():
     bad_data['names'] = bad_data['name']
     del bad_data['name']
     with pytest.raises(SwaggerDocumentationError, match='Schema key `name` was not found in the API response.'):
-        tester.test_dict(schema=schema, data=bad_data, parent='placeholder')
+        tester.test_dict(schema=schema, data=bad_data, reference='placeholder')
 
 
 def test_response_key_not_in_schema():
@@ -67,7 +67,7 @@ def test_response_key_not_in_schema():
     bad_schema['properties']['names'] = bad_schema['properties']['name']
     del bad_schema['properties']['name']
     with pytest.raises(SwaggerDocumentationError, match='Response key `name` not found in the OpenAPI schema.'):
-        tester.test_dict(schema=bad_schema, data=data, parent='placeholder')
+        tester.test_dict(schema=bad_schema, data=data, reference='placeholder')
 
 
 def test_call_dict_from_dict():
@@ -76,7 +76,7 @@ def test_call_dict_from_dict():
     """
     custom_schema = {'type': 'object', 'properties': {'test': schema}}
     custom_data = {'test': data}
-    assert tester.test_dict(schema=custom_schema, data=custom_data, parent='placeholder') is None
+    assert tester.test_dict(schema=custom_schema, data=custom_data, reference='placeholder') is None
 
 
 def test_call_list_from_dict():
@@ -93,7 +93,7 @@ def test_call_list_from_dict():
         },
     }
     custom_data = {'list': []}
-    assert tester.test_dict(schema=custom_schema, data=custom_data, parent='placeholder') is None
+    assert tester.test_dict(schema=custom_schema, data=custom_data, reference='placeholder') is None
 
 
 def test_bad_type():
@@ -113,4 +113,4 @@ def test_bad_type():
     with pytest.raises(
         Exception, match='Schema item has an invalid \`type\` attribute. The type `rarray` is not supported'
     ):
-        assert tester.test_dict(schema=custom_schema, data=custom_data, parent='placeholder') is None
+        assert tester.test_dict(schema=custom_schema, data=custom_data, reference='placeholder') is None