+++ tests/test_utils.py
@@ -45,22 +45,21 @@ def test_invalid_methods_raise():
 from django_swagger_tester.utils import unpack_response
 
 
-class MockResponse:
-    status_code = 200
-
-    @staticmethod
-    def json():
-        return {'test': 'test'}
-
-
-response = MockResponse()
-response.status_code = 200
-
-
 def test_successful_unpack():
     """
     This should run without errors.
     """
+
+    class MockResponse:
+        status_code = 200
+
+        @staticmethod
+        def json():
+            return {'test': 'test'}
+
+    response = MockResponse()
+    response.status_code = 200
+
     data, status_code = unpack_response(response)
     assert data == {'test': 'test'}
     assert status_code == 200
@@ -71,11 +70,15 @@ def test_unsuccesful_unpack():
     Verify that the appropriate error is raised when unpack fails.
     """
 
-    def _raise_exception():
-        raise Exception('test')
+    class NonJsonMockResponse:
+        status_code = 204
 
-    response.json = _raise_exception
-    with pytest.raises(Exception, match='Unable to unpack response object. Error: test'):
+    response = NonJsonMockResponse()
+
+    with pytest.raises(
+        Exception,
+        match='Response does not contain a JSON-formatted response and cannot be tested against a response schema.',
+    ):
         unpack_response(response)